+++ src/Kernels/DotBoxD.Kernels.Interpreter/ExpressionEvaluator.cs
@@ -73,6 +73,93 @@ internal sealed partial class ExpressionEvaluator
         return false;
     }
 
+    public bool TryEvaluateInt64(Expression expression, InterpreterFrame frame, out long value)
+    {
+        if (!_options.EnableDebugTrace)
+        {
+            if (I64ExpressionEvaluator.CanEvaluate(expression, frame))
+            {
+                value = I64ExpressionEvaluator.Evaluate(expression, frame, _context);
+                return true;
+            }
+
+            if (TryEvaluateInt32ToInt64(expression, frame, out value))
+            {
+                return true;
+            }
+        }
+
+        value = 0;
+        return false;
+    }
+
+    public bool TryEvaluateDouble(Expression expression, InterpreterFrame frame, out double value)
+    {
+        if (!_options.EnableDebugTrace)
+        {
+            if (F64ExpressionEvaluator.CanEvaluate(expression, frame))
+            {
+                value = F64ExpressionEvaluator.Evaluate(expression, frame, _context);
+                return true;
+            }
+
+            if (TryEvaluateToDouble(expression, frame, out value))
+            {
+                return true;
+            }
+        }
+
+        value = 0;
+        return false;
+    }
+
+    private bool TryEvaluateInt32ToInt64(
+        Expression expression,
+        InterpreterFrame frame,
+        out long value)
+    {
+        if (expression is CallExpression { Name: "numeric.toI64", Arguments.Count: 1 } call &&
+            I32ExpressionEvaluator.CanEvaluate(call.Arguments[0], frame, _interpreter))
+        {
+            _context.ChargeFuel(1);
+            value = I32ExpressionEvaluator.Evaluate(call.Arguments[0], frame, _context, _interpreter);
+            return true;
+        }
+
+        value = 0;
+        return false;
+    }
+
+    private bool TryEvaluateToDouble(
+        Expression expression,
+        InterpreterFrame frame,
+        out double value)
+    {
+        if (expression is not CallExpression { Name: "numeric.toF64", Arguments.Count: 1 } call)
+        {
+            value = 0;
+            return false;
+        }
+
+        var operand = call.Arguments[0];
+        if (I32ExpressionEvaluator.CanEvaluate(operand, frame, _interpreter))
+        {
+            _context.ChargeFuel(1);
+            value = I32ExpressionEvaluator.Evaluate(operand, frame, _context, _interpreter);
+            return true;
+        }
+
+        if (I64ExpressionEvaluator.CanEvaluate(operand, frame))
+        {
+            _context.ChargeFuel(1);
+            value = I64ExpressionEvaluator.Evaluate(operand, frame, _context);
+            return true;
+        }
+
+        value = 0;
+        return false;
+    }
+
     public ValueTask<SandboxValue> InvokeFunctionAsync(SandboxFunction function, LocalFunctionArguments args)
         => _interpreter.InvokeFunctionAsync(function, args);